Refactor code for improved readability and maintainability

+++ intetest/14_div.c
@@ -1,11 +1,11 @@
 #include<stdio.h>
 #include<stdbool.h>
 int a, b;
-double c;
+float c;
 int main() {
 a = 10;
 b = 5;
 c = a / b;
-printf("%.6lf",c);
+printf("%f",c);
 return 0;
 }
